Move helper method of base catalog class.
